@@ -26,7 +26,7 @@
         }
     },
     "require": {
-        "php": "^5.6 || ^7.0",
+        "php": "^7.3",
         "laminas-api-tools/api-tools-configuration": "^1.2",
         "laminas-api-tools/api-tools-content-negotiation": "^1.2.1",
         "laminas-api-tools/api-tools-provider": "^1.2",
@@ -34,14 +34,14 @@
         "laminas/laminas-modulemanager": "^2.7.2",
         "laminas/laminas-mvc": "^2.7.15 || ^3.0.4",
         "laminas/laminas-servicemanager": "^2.7.6 || ^3.1",
-        "laminas/laminas-view": "^2.8.1",
-        "laminas/laminas-zendframework-bridge": "^1.0",
+        "laminas/laminas-view": "^2.11.3",
+        "laminas/laminas-zendframework-bridge": "^1.1",
         "michelf/php-markdown": "^1.5"
     },
     "require-dev": {
-        "laminas/laminas-coding-standard": "~1.0.0",
-        "laminas/laminas-stdlib": "^2.7.7 || ^3.0.1",
-        "phpunit/phpunit": "^5.7.27 || ^6.5.8 || ^7.1.5"
+        "laminas/laminas-coding-standard": "~2.2.0",
+        "laminas/laminas-stdlib": "^2.7.8 || ^3.0.1",
+        "phpunit/phpunit": "^7.5.15"
     },
     "autoload": {
         "psr-4": {
