@@ -1,17 +1,10 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/api-tools-documentation for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/api-tools-documentation/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/api-tools-documentation/blob/master/LICENSE.md New BSD License
- */
-
 namespace Laminas\ApiTools\Documentation;
-
-use Laminas\ApiTools\Configuration\ModuleUtils;
 
 class Module
 {
+    /** @return array */
     public function getConfig()
     {
         return include __DIR__ . '/../config/module.config.php';
