@@ -1,21 +1,20 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/api-tools-documentation for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/api-tools-documentation/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/api-tools-documentation/blob/master/LICENSE.md New BSD License
- */
-
 namespace Laminas\ApiTools\Documentation\View;
 
 use Laminas\ApiTools\Documentation\Operation;
 use Laminas\View\Helper\AbstractHelper;
 
+use function array_map;
+use function implode;
+use function sprintf;
+
 class AgStatusCodes extends AbstractHelper
 {
+    /** @return string */
     public function __invoke(Operation $operation)
     {
-        $view = $this->getView();
+        $view        = $this->getView();
         $statusCodes = array_map(function ($status) use ($view) {
             return sprintf(
                 '<li class="list-group-item"><strong>%s:</strong> %s</li>',
