@@ -1,66 +1,73 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/api-tools-documentation for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/api-tools-documentation/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/api-tools-documentation/blob/master/LICENSE.md New BSD License
- */
-
 namespace LaminasTest\ApiTools\Documentation;
 
 use Laminas\ApiTools\Configuration\ModuleUtils;
+use Laminas\ApiTools\Documentation\Api;
 use Laminas\ApiTools\Documentation\ApiFactory;
+use Laminas\ApiTools\Documentation\Field;
+use Laminas\ApiTools\Documentation\Operation;
 use Laminas\ApiTools\Documentation\Service;
 use Laminas\ApiTools\Provider\ApiToolsProviderInterface;
 use Laminas\ModuleManager\ModuleManager;
 use PHPUnit\Framework\TestCase;
 
+use function array_filter;
+use function array_key_exists;
+use function array_shift;
+use function in_array;
+use function is_array;
+use function sprintf;
+use function var_export;
+
 class ApiFactoryTest extends TestCase
 {
-    /**
-     * @var ApiFactory
-     */
+    /** @var ApiFactory */
     protected $apiFactory;
 
+    /**
+     * @var array
+     * @psalm-var array<array-key, array{code:string, message:string}>
+     */
     protected $expectedStatusCodes = [
         [
-            'code' => '200',
+            'code'    => '200',
             'message' => 'OK',
         ],
         [
-            'code' => '201',
+            'code'    => '201',
             'message' => 'Created',
         ],
         [
-            'code' => '204',
+            'code'    => '204',
             'message' => 'No Content',
         ],
         [
-            'code' => '400',
+            'code'    => '400',
             'message' => 'Client Error',
         ],
         [
-            'code' => '401',
+            'code'    => '401',
             'message' => 'Unauthorized',
         ],
         [
-            'code' => '403',
+            'code'    => '403',
             'message' => 'Forbidden',
         ],
         [
-            'code' => '404',
+            'code'    => '404',
             'message' => 'Not Found',
         ],
         [
-            'code' => '406',
+            'code'    => '406',
             'message' => 'Not Acceptable',
         ],
         [
-            'code' => '415',
+            'code'    => '415',
             'message' => 'Unsupported Media Type',
         ],
         [
-            'code' => '422',
+            'code'    => '422',
             'message' => 'Unprocessable Entity',
         ],
     ];
@@ -78,7 +85,6 @@ class ApiFactoryTest extends TestCase
             ->getModuleConfigPath('Test')
             ->willReturn(__DIR__ . '/TestAsset/module-config/module.config.php');
 
-
         $this->apiFactory = new ApiFactory(
             $moduleManager->reveal(),
             include __DIR__ . '/TestAsset/module-config/module.config.php',
@@ -86,7 +92,11 @@ class ApiFactoryTest extends TestCase
         );
     }
 
-    public function assertContainsStatusCodes($expectedCodes, $actualCodes, $message = '')
+    /**
+     * @param int|int[]|string|string[] $expectedCodes
+     * @param int[]|string[] $actualCodes
+     */
+    public function assertContainsStatusCodes($expectedCodes, array $actualCodes, string $message = ''): void
     {
         if (! is_array($expectedCodes)) {
             $expectedCodes = [$expectedCodes];
@@ -122,8 +132,6 @@ class ApiFactoryTest extends TestCase
                 ));
             }
         }
-
-        return true;
     }
 
     public function testCreateApiList()
@@ -141,7 +149,7 @@ class ApiFactoryTest extends TestCase
     public function testCreateApi()
     {
         $api = $this->apiFactory->createApi('Test', 1);
-        $this->assertInstanceOf('Laminas\ApiTools\Documentation\Api', $api);
+        $this->assertInstanceOf(Api::class, $api);
 
         $this->assertEquals('Test', $api->getName());
         $this->assertEquals(1, $api->getVersion());
@@ -151,17 +159,17 @@ class ApiFactoryTest extends TestCase
     public function testCreateRestService()
     {
         $docConfig = include __DIR__ . '/TestAsset/module-config/documentation.config.php';
-        $api = $this->apiFactory->createApi('Test', 1);
+        $api       = $this->apiFactory->createApi('Test', 1);
 
         $service = $this->apiFactory->createService($api, 'FooBar');
-        $this->assertInstanceOf('Laminas\ApiTools\Documentation\Service', $service);
+        $this->assertInstanceOf(Service::class, $service);
 
         $this->assertEquals('FooBar', $service->getName());
         $this->assertEquals($docConfig['Test\V1\Rest\FooBar\Controller']['description'], $service->getDescription());
 
         $fields = $service->getFields('input_filter');
         $this->assertCount(5, $fields);
-        $this->assertInstanceOf('Laminas\ApiTools\Documentation\Field', $fields[0]);
+        $this->assertInstanceOf(Field::class, $fields[0]);
         $this->assertEquals('foogoober/subgoober', $fields[2]->getName());
         $this->assertEquals('foofoogoober/subgoober/subgoober', $fields[3]->getName());
 
@@ -169,7 +177,7 @@ class ApiFactoryTest extends TestCase
         $this->assertCount(2, $ops);
 
         foreach ($ops as $operation) {
-            $this->assertInstanceOf('Laminas\ApiTools\Documentation\Operation', $operation);
+            $this->assertInstanceOf(Operation::class, $operation);
             $statusCodes = $operation->getResponseStatusCodes();
             switch ($operation->getHttpMethod()) {
                 case 'GET':
@@ -193,7 +201,7 @@ class ApiFactoryTest extends TestCase
         $this->assertCount(4, $eOps);
 
         foreach ($eOps as $operation) {
-            $this->assertInstanceOf('Laminas\ApiTools\Documentation\Operation', $operation);
+            $this->assertInstanceOf(Operation::class, $operation);
             $statusCodes = $operation->getResponseStatusCodes();
             switch ($operation->getHttpMethod()) {
                 case 'GET':
@@ -222,10 +230,10 @@ class ApiFactoryTest extends TestCase
     public function testCreateRestServiceWithCollection()
     {
         $docConfig = include __DIR__ . '/TestAsset/module-config/documentation.config.php';
-        $api = $this->apiFactory->createApi('Test', 1);
+        $api       = $this->apiFactory->createApi('Test', 1);
 
         $service = $this->apiFactory->createService($api, 'FooBarCollection');
-        $this->assertInstanceOf('Laminas\ApiTools\Documentation\Service', $service);
+        $this->assertInstanceOf(Service::class, $service);
 
         $this->assertEquals('FooBarCollection', $service->getName());
         $this->assertEquals(
@@ -274,10 +282,10 @@ class ApiFactoryTest extends TestCase
     public function testCreateRpcService()
     {
         $docConfig = include __DIR__ . '/TestAsset/module-config/documentation.config.php';
-        $api = $this->apiFactory->createApi('Test', 1);
+        $api       = $this->apiFactory->createApi('Test', 1);
 
         $service = $this->apiFactory->createService($api, 'Ping');
-        $this->assertInstanceOf('Laminas\ApiTools\Documentation\Service', $service);
+        $this->assertInstanceOf(Service::class, $service);
 
         $this->assertEquals('Ping', $service->getName());
         $this->assertEquals($docConfig['Test\V1\Rpc\Ping\Controller']['description'], $service->getDescription());
@@ -286,7 +294,7 @@ class ApiFactoryTest extends TestCase
         $this->assertCount(1, $ops);
 
         foreach ($ops as $operation) {
-            $this->assertInstanceOf('Laminas\ApiTools\Documentation\Operation', $operation);
+            $this->assertInstanceOf(Operation::class, $operation);
             $statusCodes = $operation->getResponseStatusCodes();
             switch ($operation->getHttpMethod()) {
                 case 'GET':
@@ -314,7 +322,7 @@ class ApiFactoryTest extends TestCase
 
     public function testGetFieldsForEntityMethods()
     {
-        $api = $this->apiFactory->createApi('Test', 1);
+        $api     = $this->apiFactory->createApi('Test', 1);
         $service = $this->apiFactory->createService($api, 'EntityFields');
         $this->assertEquals('EntityFields', $service->getName());
         $this->assertCount(1, $service->getFields('PUT'));
