@@ -5,43 +5,39 @@ namespace LaminasTest\ApiTools\Documentation;
 use Laminas\ApiTools\Documentation\ApiFactory;
 use Laminas\ApiTools\Documentation\Controller;
 use Laminas\Mvc\MvcEvent;
+use Laminas\Router\RouteMatch;
 use Laminas\View\Helper\BasePath;
 use Laminas\View\Helper\ServerUrl;
 use Laminas\View\Model\ModelInterface;
 use PHPUnit\Framework\TestCase;
+use PHPUnit_Framework_MockObject_MockObject;
+
+use function class_exists;
 
 class ControllerTest extends TestCase
 {
-    /**
-     * @var MvcEvent
-     */
+    /** @var MvcEvent */
     private $event;
 
-    /**
-     * @var ServerUrl
-     */
+    /** @var ServerUrl */
     private $serverUrl;
 
-    /**
-     * @var BasePath
-     */
+    /** @var BasePath */
     private $basePath;
 
-    /**
-     * @var ApiFactory|\PHPUnit_Framework_MockObject_MockObject
-     */
+    /** @var ApiFactory|PHPUnit_Framework_MockObject_MockObject */
     private $apiFactory;
 
     protected function setUp()
     {
         $this->apiFactory = $this->getMockBuilder(ApiFactory::class)->disableOriginalConstructor()->getMock();
-        $this->serverUrl = new ServerUrl();
-        $this->basePath = new BasePath();
+        $this->serverUrl  = new ServerUrl();
+        $this->basePath   = new BasePath();
 
         $this->event = new MvcEvent();
 
-        if (class_exists('Laminas\Router\RouteMatch', true)) {
-            $this->event->setRouteMatch(new \Laminas\Router\RouteMatch([]));
+        if (class_exists(RouteMatch::class, true)) {
+            $this->event->setRouteMatch(new RouteMatch([]));
         } elseif (class_exists('Laminas\Mvc\Router\RouteMatch', true)) {
             $this->event->setRouteMatch(new \Laminas\Mvc\Router\RouteMatch([]));
         }
