@@ -1,75 +1,72 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/api-tools-documentation for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/api-tools-documentation/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/api-tools-documentation/blob/master/LICENSE.md New BSD License
- */
+use Laminas\InputFilter\CollectionInputFilter;
+use Laminas\InputFilter\InputFilter;
 
 return [
-    'router' => [
+    'router'                        => [
         'routes' => [
-            'test.rest.foo-bar' => [
-                'type' => 'Segment',
+            'test.rest.foo-bar'            => [
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/foo-bar[/:foo_bar_id]',
+                    'route'    => '/foo-bar[/:foo_bar_id]',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rest\\FooBar\\Controller',
                     ],
                 ],
             ],
             'test.rest.foo-bar-collection' => [
-                'type' => 'Segment',
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/foo-bar-collection[/:foo_bar_collection_id]',
+                    'route'    => '/foo-bar-collection[/:foo_bar_collection_id]',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rest\\FooBarCollection\\Controller',
                     ],
                 ],
             ],
-            'test.rest.boo-baz' => [
-                'type' => 'Segment',
+            'test.rest.boo-baz'            => [
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/boo-baz[/:boo_baz_id]',
+                    'route'    => '/boo-baz[/:boo_baz_id]',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rest\\BooBaz\\Controller',
                     ],
                 ],
             ],
-            'test.rpc.my-rpc' => [
-                'type' => 'Segment',
+            'test.rpc.my-rpc'              => [
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/my-rpc',
+                    'route'    => '/my-rpc',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rpc\\MyRpc\\Controller',
-                        'action' => 'myRpc',
+                        'action'     => 'myRpc',
                     ],
                 ],
             ],
-            'test.rpc.ping' => [
-                'type' => 'Segment',
+            'test.rpc.ping'                => [
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/ping',
+                    'route'    => '/ping',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rpc\\Ping\\Controller',
-                        'action' => 'ping',
+                        'action'     => 'ping',
                     ],
                 ],
             ],
-            'test.rest.entity-fields' => [
-                'type' => 'Segment',
+            'test.rest.entity-fields'      => [
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/entity-fields',
+                    'route'    => '/entity-fields',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rest\\EntityFields\\Controller',
-                        'action' => 'test',
+                        'action'     => 'test',
                     ],
                 ],
             ],
-            'test.rest.bands' => [
-                'type' => 'Segment',
+            'test.rest.bands'              => [
+                'type'    => 'Segment',
                 'options' => [
-                    'route' => '/bands[/:band_id]',
+                    'route'    => '/bands[/:band_id]',
                     'defaults' => [
                         'controller' => 'Test\\V1\\Rest\\Bands\\Controller',
                     ],
@@ -77,7 +74,7 @@ return [
             ],
         ],
     ],
-    'api-tools-versioning' => [
+    'api-tools-versioning'          => [
         'uri' => [
             0 => 'test.rest.foo-bar',
             1 => 'test.rest.boo-baz',
@@ -86,132 +83,131 @@ return [
             4 => 'test.rest.foo-bar-collection',
         ],
     ],
-    'service_manager' => [
+    'service_manager'               => [
         'invokables' => [
-            'Test\\V1\\Rest\\FooBar\\FooBarResource' => 'Test\\V1\\Rest\\FooBar\\FooBarResource',
+            'Test\\V1\\Rest\\FooBar\\FooBarResource'           => 'Test\\V1\\Rest\\FooBar\\FooBarResource',
             'Test\\V1\\Rest\\FooBarCollection\\FooBarResource' => 'Test\\V1\\Rest\\FooBarCollection\\FooBarResource',
-            'Test\\V1\\Rest\\BooBaz\\BooBazResource' => 'Test\\V1\\Rest\\BooBaz\\BooBazResource',
-            'Test\\V1\\Rest\\Bands\\BandsResource' => 'Test\\V1\\Rest\\Bands\\BandsResource',
+            'Test\\V1\\Rest\\BooBaz\\BooBazResource'           => 'Test\\V1\\Rest\\BooBaz\\BooBazResource',
+            'Test\\V1\\Rest\\Bands\\BandsResource'             => 'Test\\V1\\Rest\\Bands\\BandsResource',
         ],
     ],
-    'api-tools-rest' => [
-        'Test\\V1\\Rest\\FooBar\\Controller' => [
-            'listener' => 'Test\\V1\\Rest\\FooBar\\FooBarResource',
-            'route_name' => 'test.rest.foo-bar',
-            'route_identifier_name' => 'foo_bar_id',
-            'collection_name' => 'foo_bar',
-            'entity_http_methods' => [
+    'api-tools-rest'                => [
+        'Test\\V1\\Rest\\FooBar\\Controller'           => [
+            'listener'                   => 'Test\\V1\\Rest\\FooBar\\FooBarResource',
+            'route_name'                 => 'test.rest.foo-bar',
+            'route_identifier_name'      => 'foo_bar_id',
+            'collection_name'            => 'foo_bar',
+            'entity_http_methods'        => [
                 0 => 'GET',
                 1 => 'PATCH',
                 2 => 'PUT',
                 3 => 'DELETE',
             ],
-            'collection_http_methods' => [
+            'collection_http_methods'    => [
                 0 => 'GET',
                 1 => 'POST',
             ],
             'collection_query_whitelist' => [],
-            'page_size' => 25,
-            'page_size_param' => null,
-            'entity_class' => 'Test\\V1\\Rest\\FooBar\\FooBarEntity',
-            'collection_class' => 'Test\\V1\\Rest\\FooBar\\FooBarCollection',
-            'service_name' => 'FooBar',
+            'page_size'                  => 25,
+            'page_size_param'            => null,
+            'entity_class'               => 'Test\\V1\\Rest\\FooBar\\FooBarEntity',
+            'collection_class'           => 'Test\\V1\\Rest\\FooBar\\FooBarCollection',
+            'service_name'               => 'FooBar',
         ],
         'Test\\V1\\Rest\\FooBarCollection\\Controller' => [
-            'listener' => 'Test\\V1\\Rest\\FooBarCollection\\FooBarResource',
-            'route_name' => 'test.rest.foo-bar-collection',
-            'route_identifier_name' => 'foo_bar_collectio_id',
-            'collection_name' => 'foo_bar_collection',
-            'entity_http_methods' => [
+            'listener'                   => 'Test\\V1\\Rest\\FooBarCollection\\FooBarResource',
+            'route_name'                 => 'test.rest.foo-bar-collection',
+            'route_identifier_name'      => 'foo_bar_collectio_id',
+            'collection_name'            => 'foo_bar_collection',
+            'entity_http_methods'        => [
                 0 => 'GET',
                 1 => 'PATCH',
                 2 => 'PUT',
                 3 => 'DELETE',
             ],
-            'collection_http_methods' => [
+            'collection_http_methods'    => [
                 0 => 'GET',
                 1 => 'POST',
             ],
             'collection_query_whitelist' => [],
-            'page_size' => 25,
-            'page_size_param' => null,
-            'entity_class' => 'Test\\V1\\Rest\\FooBarCollection\\FooBarEntity',
-            'collection_class' => 'Test\\V1\\Rest\\FooBarCollection\\FooBarCollection',
-            'service_name' => 'FooBarCollection',
+            'page_size'                  => 25,
+            'page_size_param'            => null,
+            'entity_class'               => 'Test\\V1\\Rest\\FooBarCollection\\FooBarEntity',
+            'collection_class'           => 'Test\\V1\\Rest\\FooBarCollection\\FooBarCollection',
+            'service_name'               => 'FooBarCollection',
         ],
-        'Test\\V1\\Rest\\BooBaz\\Controller' => [
-            'listener' => 'Test\\V1\\Rest\\BooBaz\\BooBazResource',
-            'route_name' => 'test.rest.boo-baz',
-            'route_identifier_name' => 'boo_baz_id',
-            'collection_name' => 'boo_baz',
-            'entity_http_methods' => [
+        'Test\\V1\\Rest\\BooBaz\\Controller'           => [
+            'listener'                   => 'Test\\V1\\Rest\\BooBaz\\BooBazResource',
+            'route_name'                 => 'test.rest.boo-baz',
+            'route_identifier_name'      => 'boo_baz_id',
+            'collection_name'            => 'boo_baz',
+            'entity_http_methods'        => [
                 0 => 'GET',
                 1 => 'PATCH',
                 2 => 'PUT',
                 3 => 'DELETE',
             ],
-            'collection_http_methods' => [
+            'collection_http_methods'    => [
                 0 => 'GET',
                 1 => 'POST',
             ],
             'collection_query_whitelist' => [],
-            'page_size' => 25,
-            'page_size_param' => null,
-            'entity_class' => 'Test\\V1\\Rest\\BooBaz\\BooBazEntity',
-            'collection_class' => 'Test\\V1\\Rest\\BooBaz\\BooBazCollection',
-            'service_name' => 'BooBaz',
+            'page_size'                  => 25,
+            'page_size_param'            => null,
+            'entity_class'               => 'Test\\V1\\Rest\\BooBaz\\BooBazEntity',
+            'collection_class'           => 'Test\\V1\\Rest\\BooBaz\\BooBazCollection',
+            'service_name'               => 'BooBaz',
         ],
-        'Test\\V1\\Rest\\EntityFields\\Controller' => [
-            'listener' => 'Test\\V1\\Rest\\EntityFields\\EntityFieldsResource',
-            'route_name' => 'test.rest.entity-fields',
-            'route_identifier_name' => 'id',
-            'collection_name' => 'entity_fields',
-            'entity_http_methods' => [
+        'Test\\V1\\Rest\\EntityFields\\Controller'     => [
+            'listener'                   => 'Test\\V1\\Rest\\EntityFields\\EntityFieldsResource',
+            'route_name'                 => 'test.rest.entity-fields',
+            'route_identifier_name'      => 'id',
+            'collection_name'            => 'entity_fields',
+            'entity_http_methods'        => [
                 0 => 'PUT',
             ],
-            'collection_http_methods' => [
-            ],
+            'collection_http_methods'    => [],
             'collection_query_whitelist' => [],
-            'page_size' => 25,
-            'page_size_param' => null,
-            'entity_class' => 'Test\\V1\\Rest\\EntityFields\\EntityFieldsEntity',
-            'collection_class' => 'Test\\V1\\Rest\\EntityFields\\EntityFieldsCollection',
-            'service_name' => 'EntityFields',
+            'page_size'                  => 25,
+            'page_size_param'            => null,
+            'entity_class'               => 'Test\\V1\\Rest\\EntityFields\\EntityFieldsEntity',
+            'collection_class'           => 'Test\\V1\\Rest\\EntityFields\\EntityFieldsCollection',
+            'service_name'               => 'EntityFields',
         ],
-        'Test\\V1\\Rest\\Bands\\Controller' => [
-            'listener' => 'Test\\V1\\Rest\\Bands\\BandsResource',
-            'route_name' => 'test.rest.bands',
-            'route_identifier_name' => 'artist_id',
-            'collection_name' => 'foo_bar',
-            'entity_http_methods' => [
+        'Test\\V1\\Rest\\Bands\\Controller'            => [
+            'listener'                   => 'Test\\V1\\Rest\\Bands\\BandsResource',
+            'route_name'                 => 'test.rest.bands',
+            'route_identifier_name'      => 'artist_id',
+            'collection_name'            => 'foo_bar',
+            'entity_http_methods'        => [
                 0 => 'GET',
                 1 => 'PATCH',
                 2 => 'PUT',
                 3 => 'DELETE',
             ],
-            'collection_http_methods' => [
+            'collection_http_methods'    => [
                 0 => 'GET',
                 1 => 'POST',
             ],
             'collection_query_whitelist' => [],
-            'page_size' => 25,
-            'page_size_param' => null,
-            'entity_class' => 'Test\\V1\\Rest\\Bands\\ArtistEntity',
-            'collection_class' => 'Test\\V1\\Rest\\Bands\\ArtistCollection',
-            'service_name' => 'Bands',
+            'page_size'                  => 25,
+            'page_size_param'            => null,
+            'entity_class'               => 'Test\\V1\\Rest\\Bands\\ArtistEntity',
+            'collection_class'           => 'Test\\V1\\Rest\\Bands\\ArtistCollection',
+            'service_name'               => 'Bands',
         ],
     ],
     'api-tools-content-negotiation' => [
-        'controllers' => [
-            'Test\\V1\\Rest\\FooBar\\Controller' => 'HalJson',
+        'controllers'            => [
+            'Test\\V1\\Rest\\FooBar\\Controller'           => 'HalJson',
             'Test\\V1\\Rest\\FooBarCollection\\Controller' => 'HalJson',
-            'Test\\V1\\Rest\\BooBaz\\Controller' => 'HalJson',
-            'Test\\V1\\Rpc\\MyRpc\\Controller' => 'Json',
-            'Test\\V1\\Rpc\\Ping\\Controller' => 'Json',
-            'Test\\V1\\Rest\\Bands\\Controller' => 'HalJson',
+            'Test\\V1\\Rest\\BooBaz\\Controller'           => 'HalJson',
+            'Test\\V1\\Rpc\\MyRpc\\Controller'             => 'Json',
+            'Test\\V1\\Rpc\\Ping\\Controller'              => 'Json',
+            'Test\\V1\\Rest\\Bands\\Controller'            => 'HalJson',
         ],
-        'accept_whitelist' => [
-            'Test\\V1\\Rest\\FooBar\\Controller' => [
+        'accept_whitelist'       => [
+            'Test\\V1\\Rest\\FooBar\\Controller'           => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/hal+json',
                 2 => 'application/json',
@@ -221,34 +217,34 @@ return [
                 1 => 'application/hal+json',
                 2 => 'application/json',
             ],
-            'Test\\V1\\Rest\\BooBaz\\Controller' => [
+            'Test\\V1\\Rest\\BooBaz\\Controller'           => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/hal+json',
                 2 => 'application/json',
             ],
-            'Test\\V1\\Rpc\\MyRpc\\Controller' => [
+            'Test\\V1\\Rpc\\MyRpc\\Controller'             => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
                 2 => 'application/*+json',
             ],
-            'Test\\V1\\Rpc\\Ping\\Controller' => [
+            'Test\\V1\\Rpc\\Ping\\Controller'              => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
                 2 => 'application/*+json',
             ],
-            'Test\\V1\\Rest\\EntityFields\\Controller' => [
+            'Test\\V1\\Rest\\EntityFields\\Controller'     => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
                 2 => 'application/*+json',
             ],
-            'Test\\V1\\Rest\\Bands\\Controller' => [
+            'Test\\V1\\Rest\\Bands\\Controller'            => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/hal+json',
                 2 => 'application/json',
             ],
         ],
         'content_type_whitelist' => [
-            'Test\\V1\\Rest\\FooBar\\Controller' => [
+            'Test\\V1\\Rest\\FooBar\\Controller'           => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
@@ -256,287 +252,285 @@ return [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
-            'Test\\V1\\Rest\\BooBaz\\Controller' => [
+            'Test\\V1\\Rest\\BooBaz\\Controller'           => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
-            'Test\\V1\\Rpc\\MyRpc\\Controller' => [
+            'Test\\V1\\Rpc\\MyRpc\\Controller'             => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
-            'Test\\V1\\Rpc\\Ping\\Controller' => [
+            'Test\\V1\\Rpc\\Ping\\Controller'              => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
-            'Test\\V1\\Rpc\\EntityFields\\Controller' => [
+            'Test\\V1\\Rpc\\EntityFields\\Controller'      => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
-            'Test\\V1\\Rest\\Bands\\Controller' => [
+            'Test\\V1\\Rest\\Bands\\Controller'            => [
                 0 => 'application/vnd.test.v1+json',
                 1 => 'application/json',
             ],
         ],
     ],
-    'api-tools-hal' => [
+    'api-tools-hal'                 => [
         'metadata_map' => [
-            'Test\\V1\\Rest\\FooBar\\FooBarEntity' => [
+            'Test\\V1\\Rest\\FooBar\\FooBarEntity'                 => [
                 'entity_identifier_name' => 'id',
-                'route_name' => 'test.rest.foo-bar',
-                'route_identifier_name' => 'foo_bar_id',
-                'hydrator' => 'Laminas\\Hydrator\\ArraySerializable',
+                'route_name'             => 'test.rest.foo-bar',
+                'route_identifier_name'  => 'foo_bar_id',
+                'hydrator'               => 'Laminas\\Hydrator\\ArraySerializable',
             ],
-            'Test\\V1\\Rest\\FooBar\\FooBarCollection' => [
+            'Test\\V1\\Rest\\FooBar\\FooBarCollection'             => [
                 'entity_identifier_name' => 'id',
-                'route_name' => 'test.rest.foo-bar',
-                'route_identifier_name' => 'foo_bar_id',
-                'is_collection' => true,
+                'route_name'             => 'test.rest.foo-bar',
+                'route_identifier_name'  => 'foo_bar_id',
+                'is_collection'          => true,
             ],
-            'Test\\V1\\Rest\\BooBaz\\BooBazEntity' => [
+            'Test\\V1\\Rest\\BooBaz\\BooBazEntity'                 => [
                 'entity_identifier_name' => 'id',
-                'route_name' => 'test.rest.boo-baz',
-                'route_identifier_name' => 'boo_baz_id',
-                'hydrator' => 'Laminas\\Hydrator\\ArraySerializable',
+                'route_name'             => 'test.rest.boo-baz',
+                'route_identifier_name'  => 'boo_baz_id',
+                'hydrator'               => 'Laminas\\Hydrator\\ArraySerializable',
             ],
-            'Test\\V1\\Rest\\BooBaz\\BooBazCollection' => [
+            'Test\\V1\\Rest\\BooBaz\\BooBazCollection'             => [
                 'entity_identifier_name' => 'id',
-                'route_name' => 'test.rest.boo-baz',
-                'route_identifier_name' => 'boo_baz_id',
-                'is_collection' => true,
+                'route_name'             => 'test.rest.boo-baz',
+                'route_identifier_name'  => 'boo_baz_id',
+                'is_collection'          => true,
             ],
             'Test\\V1\\Rest\\EntityFields\\EntityFieldsCollection' => [
                 'entity_identifier_name' => 'id',
-                'route_name' => 'test.rest.entity-fields',
-                'route_identifier_name' => 'id',
-                'is_collection' => true,
+                'route_name'             => 'test.rest.entity-fields',
+                'route_identifier_name'  => 'id',
+                'is_collection'          => true,
             ],
-            'Test\\V1\\Rest\\Bands\\ArtistEntity' => [
+            'Test\\V1\\Rest\\Bands\\ArtistEntity'                  => [
                 'entity_identifier_name' => 'id',
-                'route_name' => 'test.rest.bands',
-                'route_identifier_name' => 'artist_id',
-                'hydrator' => 'Laminas\\Hydrator\\ArraySerializable',
+                'route_name'             => 'test.rest.bands',
+                'route_identifier_name'  => 'artist_id',
+                'hydrator'               => 'Laminas\\Hydrator\\ArraySerializable',
             ],
         ],
     ],
-    'controllers' => [
+    'controllers'                   => [
         'invokables' => [
             'Test\\V1\\Rpc\\MyRpc\\Controller' => 'Test\\V1\\Rpc\\MyRpc\\MyRpcController',
-            'Test\\V1\\Rpc\\Ping\\Controller' => 'Test\\V1\\Rpc\\Ping\\PingController',
+            'Test\\V1\\Rpc\\Ping\\Controller'  => 'Test\\V1\\Rpc\\Ping\\PingController',
         ],
     ],
-    'api-tools-rpc' => [
+    'api-tools-rpc'                 => [
         'Test\\V1\\Rpc\\MyRpc\\Controller' => [
             'service_name' => 'MyRpc',
             'http_methods' => [
                 0 => 'GET',
             ],
-            'route_name' => 'test.rpc.my-rpc',
+            'route_name'   => 'test.rpc.my-rpc',
         ],
-        'Test\\V1\\Rpc\\Ping\\Controller' => [
+        'Test\\V1\\Rpc\\Ping\\Controller'  => [
             'service_name' => 'Ping',
             'http_methods' => [
                 0 => 'GET',
             ],
-            'route_name' => 'test.rpc.ping',
+            'route_name'   => 'test.rpc.ping',
         ],
     ],
-    'api-tools-content-validation' => [
-        'Test\\V1\\Rest\\FooBar\\Controller' => [
+    'api-tools-content-validation'  => [
+        'Test\\V1\\Rest\\FooBar\\Controller'           => [
             'input_filter' => 'Test\\V1\\Rest\\FooBar\\Validator',
         ],
         'Test\\V1\\Rest\\FooBarCollection\\Controller' => [
             'input_filter' => 'Test\\V1\\Rest\\FooBarCollection\\Validator',
         ],
-        'Test\\V1\\Rest\\EntityFields\\Controller' => [
+        'Test\\V1\\Rest\\EntityFields\\Controller'     => [
             'input_filter' => 'Test\\V1\\Rest\\EntityFields\\Validator',
-            'PUT' => 'Test\\V1\\Rest\\EntityFields\\Validator\\Put',
+            'PUT'          => 'Test\\V1\\Rest\\EntityFields\\Validator\\Put',
         ],
-        'Test\\V1\\Rest\\Bands\\Controller' => [
+        'Test\\V1\\Rest\\Bands\\Controller'            => [
             'input_filter' => 'Test\\V1\\Rest\\Bands\\Validator',
         ],
     ],
-    'input_filter_specs' => [
-        'Test\\V1\\Rest\\FooBar\\Validator' => [
-            0 => [
-                'name' => 'goober',
-                'required' => true,
-                'filters' => [],
-                'validators' => [],
+    'input_filter_specs'            => [
+        'Test\\V1\\Rest\\FooBar\\Validator'            => [
+            0                => [
+                'name'        => 'goober',
+                'required'    => true,
+                'filters'     => [],
+                'validators'  => [],
                 'description' => 'This is the description for goober.',
             ],
-            1 => [
-                'name' => 'bergoo',
-                'required' => true,
-                'filters' => [],
+            1                => [
+                'name'       => 'bergoo',
+                'required'   => true,
+                'filters'    => [],
                 'validators' => [],
             ],
-            'foogoober' => [
-                'type' => 'Laminas\InputFilter\InputFilter',
+            'foogoober'      => [
+                'type'      => InputFilter::class,
                 'subgoober' => [
-                    'name' => 'subgoober',
-                    'required' => true,
-                    'filters' => [],
+                    'name'       => 'subgoober',
+                    'required'   => true,
+                    'filters'    => [],
                     'validators' => [],
-                ]
+                ],
             ],
-            'foofoogoober' => [
-                'type' => 'Laminas\InputFilter\InputFilter',
+            'foofoogoober'   => [
+                'type'      => InputFilter::class,
                 'subgoober' => [
-                    'type' => 'Laminas\InputFilter\InputFilter',
+                    'type'      => InputFilter::class,
                     'subgoober' => [
-                        'name' => 'subgoober',
-                        'required' => true,
-                        'filters' => [],
+                        'name'       => 'subgoober',
+                        'required'   => true,
+                        'filters'    => [],
                         'validators' => [],
-                    ]
-                ]
+                    ],
+                ],
             ],
             'companyDetails' => [
-                'type' => 'Laminas\\InputFilter\\InputFilter',
-                'name' => [
-                    'name' => 'name',
-                    'required' => true,
-                    'validators' => [
-                    ],
-                    'description' => '',
-                    'allow_empty' => false,
+                'type'        => InputFilter::class,
+                'name'        => [
+                    'name'              => 'name',
+                    'required'          => true,
+                    'validators'        => [],
+                    'description'       => '',
+                    'allow_empty'       => false,
                     'continue_if_empty' => false,
                 ],
-                'required' => [
-                    'allow_empty' => false,
+                'required'    => [
+                    'allow_empty'       => false,
                     'continue_if_empty' => false,
                 ],
                 'description' => [
-                    'name' => 'website',
-                    'required' => false,
-                    'validators' => [
-                    ],
-                    'allow_empty' => false,
+                    'name'              => 'website',
+                    'required'          => false,
+                    'validators'        => [],
+                    'allow_empty'       => false,
                     'continue_if_empty' => false,
                 ],
             ],
         ],
-        'Test\\V1\\Rest\\FooBarCollection\\Validator' => [
-            'FooBarCollection' => [
-                'type' => Laminas\InputFilter\CollectionInputFilter::class,
-                'required' => true,
-                'count' => 1,
+        'Test\\V1\\Rest\\FooBarCollection\\Validator'  => [
+            'FooBarCollection'  => [
+                'type'         => CollectionInputFilter::class,
+                'required'     => true,
+                'count'        => 1,
                 'input_filter' => [
-                    'type' => Laminas\InputFilter\InputFilter::class,
-                    'name' => 'FooBar',
-                    'required' => true,
-                    'filters' => [],
+                    'type'       => InputFilter::class,
+                    'name'       => 'FooBar',
+                    'required'   => true,
+                    'filters'    => [],
                     'validators' => [],
                 ],
             ],
             'AnotherCollection' => [
-                'type' => 'Laminas\\InputFilter\\CollectionInputFilter',
-                'required' => true,
-                'count' => 1,
+                'type'         => CollectionInputFilter::class,
+                'required'     => true,
+                'count'        => 1,
                 'input_filter' => [
-                    'type' => Laminas\InputFilter\InputFilter::class,
-                    'name' => 'FooBar',
-                    'required' => true,
-                    'filters' => [],
+                    'type'       => InputFilter::class,
+                    'name'       => 'FooBar',
+                    'required'   => true,
+                    'filters'    => [],
                     'validators' => [],
                 ],
             ],
         ],
-        'Test\\V1\\Rest\\EntityFields\\Validator' => [
+        'Test\\V1\\Rest\\EntityFields\\Validator'      => [
             0 => [
-                'required' => true,
-                'validators' => [],
-                'filters' => [],
-                'name' => 'test',
+                'required'    => true,
+                'validators'  => [],
+                'filters'     => [],
+                'name'        => 'test',
                 'description' => 'test',
             ],
         ],
         'Test\\V1\\Rest\\EntityFields\\Validator\\Put' => [
             0 => [
-                'required' => true,
-                'validators' => [],
-                'filters' => [],
-                'name' => 'test_put',
+                'required'    => true,
+                'validators'  => [],
+                'filters'     => [],
+                'name'        => 'test_put',
                 'description' => 'test_put',
             ],
         ],
-        'Test\\V1\\Rest\\Bands\\Validator' => [
+        'Test\\V1\\Rest\\Bands\\Validator'             => [
             [
-                'name' => 'name',
-                'required' => true,
+                'name'        => 'name',
+                'required'    => true,
                 'description' => 'The name of the Band.',
             ],
-            'artists' => [
-                'type' => Laminas\InputFilter\CollectionInputFilter::class,
+            'artists'     => [
+                'type'         => CollectionInputFilter::class,
                 'input_filter' => [
-                    'type' => \Laminas\InputFilter\InputFilter::class,
+                    'type'       => InputFilter::class,
                     'first_name' => [
-                        'name' => 'first_name',
-                        'required' => true,
+                        'name'        => 'first_name',
+                        'required'    => true,
                         'description' => 'The Artist\'s first name.',
                     ],
-                    'last_name' => [
-                        'name' => 'last_name',
-                        'required' => true,
+                    'last_name'  => [
+                        'name'        => 'last_name',
+                        'required'    => true,
                         'description' => 'The Artist\'s last name.',
                     ],
                 ],
             ],
             'debut_album' => [
-                'type' => \Laminas\InputFilter\InputFilter::class,
-                'title' => [
-                    'name' => 'title',
-                    'required' => true,
+                'type'         => InputFilter::class,
+                'title'        => [
+                    'name'        => 'title',
+                    'required'    => true,
                     'description' => 'Album title.',
                 ],
                 'release_date' => [
-                    'name' => 'release_date',
-                    'required' => true,
+                    'name'        => 'release_date',
+                    'required'    => true,
                     'description' => 'Album release date.',
                 ],
-                'tracks' => [
-                    'type' => Laminas\InputFilter\CollectionInputFilter::class,
+                'tracks'       => [
+                    'type'         => CollectionInputFilter::class,
                     'input_filter' => [
-                        'type' => \Laminas\InputFilter\InputFilter::class,
+                        'type'   => InputFilter::class,
                         'number' => [
-                            'name' => 'number',
-                            'required' => true,
+                            'name'        => 'number',
+                            'required'    => true,
                             'description' => 'Track number.',
                         ],
-                        'title' => [
-                            'name' => 'title',
-                            'required' => true,
+                        'title'  => [
+                            'name'        => 'title',
+                            'required'    => true,
                             'description' => 'Track title.',
                         ],
                     ],
                 ],
             ],
-            'albums' => [
-                'type' => Laminas\InputFilter\CollectionInputFilter::class,
+            'albums'      => [
+                'type'         => CollectionInputFilter::class,
                 'input_filter' => [
-                    'type' => \Laminas\InputFilter\InputFilter::class,
-                    'title' => [
-                        'name' => 'title',
-                        'required' => true,
+                    'type'         => InputFilter::class,
+                    'title'        => [
+                        'name'        => 'title',
+                        'required'    => true,
                         'description' => 'Album title.',
                     ],
                     'release_date' => [
-                        'name' => 'release_date',
-                        'required' => true,
+                        'name'        => 'release_date',
+                        'required'    => true,
                         'description' => 'Album release date.',
                     ],
-                    'tracks' => [
-                        'type' => Laminas\InputFilter\CollectionInputFilter::class,
+                    'tracks'       => [
+                        'type'         => CollectionInputFilter::class,
                         'input_filter' => [
-                            'type' => \Laminas\InputFilter\InputFilter::class,
+                            'type'   => InputFilter::class,
                             'number' => [
-                                'name' => 'number',
-                                'required' => true,
+                                'name'        => 'number',
+                                'required'    => true,
                                 'description' => 'Track number.',
                             ],
-                            'title' => [
-                                'name' => 'title',
-                                'required' => true,
+                            'title'  => [
+                                'name'        => 'title',
+                                'required'    => true,
                                 'description' => 'Track title.',
                             ],
                         ],
@@ -545,16 +539,16 @@ return [
             ],
         ],
     ],
-    'api-tools-mvc-auth' => [
+    'api-tools-mvc-auth'            => [
         'authentication' => [
             'http' => [
-                'realm' => 'api',
+                'realm'    => 'api',
                 'htpasswd' => __DIR__ . '/htpasswd',
             ],
         ],
-        'authorization' => [
-            'Test\V1\Rest\FooBar\Controller' => [
-                'entity' => [
+        'authorization'  => [
+            'Test\V1\Rest\FooBar\Controller'           => [
+                'entity'     => [
                     'DELETE' => true,
                     'GET'    => false,
                     'PATCH'  => true,
@@ -570,7 +564,7 @@ return [
                 ],
             ],
             'Test\V1\Rest\FooBarCollection\Controller' => [
-                'entity' => [
+                'entity'     => [
                     'DELETE' => true,
                     'GET'    => false,
                     'PATCH'  => true,
@@ -585,8 +579,8 @@ return [
                     'PUT'    => false,
                 ],
             ],
-            'Test\V1\Rest\BooBaz\Controller' => [
-                'entity' => [
+            'Test\V1\Rest\BooBaz\Controller'           => [
+                'entity'     => [
                     'DELETE' => true,
                     'GET'    => false,
                     'PATCH'  => true,
@@ -601,7 +595,7 @@ return [
                     'PUT'    => false,
                 ],
             ],
-            'Test\V1\Rpc\MyRpc\Controller' => [
+            'Test\V1\Rpc\MyRpc\Controller'             => [
                 'actions' => [
                     'myRpc' => [
                         'DELETE' => false,
@@ -612,7 +606,7 @@ return [
                     ],
                 ],
             ],
-            'Test\V1\Rpc\Ping\Controller' => [
+            'Test\V1\Rpc\Ping\Controller'              => [
                 'actions' => [
                     'ping' => [
                         'DELETE' => false,
